아이템 리스트 구조 변경
